made some small modification

+++ Question02.js
@@ -2,5 +2,5 @@
 
 const readLine = require('readline-sync');
 let num1 = parseInt(readLine.question("Enter the number 1\n"));
-let num2 = parseInt(readLine.question("Enter the number 2\n"));
+let num2 = parseFloat(readLine.question("Enter the number 2\n"));
 console.log(`The sum of the two numbers are ${num1+num2}.`)
